out/templates: fix potential race condition
A new template syntax was introduced by this fix, so all layout
templates will need to be updated upon rebuild with this version.

In layout templates, {{ content }} should be replaced with
{{ $.Content }}.

In all templates, occurrences of dot {{ . }} should be replaced with
{{ .Data }}.

During concurrent template renderings, panics would often occur in
various places originating from compress/gzip and compress/flate. I'm
not entirely sure what the mechanism was, but it appeared to be fixed by
addressing a non-threadsafe way of mutating the template funcs attached
to layouts.

Previously, inner content was bound to the layout template by closing
over the content and assigning that closure as a template func to the
layout. However, the layout is a globally shared template, and
concurrent renderings would potentially wreak havoc by mutating the
global layout's func map with an unrelated closure.

The fix was to render templates with a package-provided context object
instead of implicitly using a closed-over context. The template syntax
is slightly less pretty but it doesn't have the race condition.

I probably need to figure out how to write a test case for this, but for
the time being, I have run a hand-wavey test that seems to confirm the
fix. go test -race also agrees, though I don't know if it would have
caught anything before.